deleteItem method added

+++ src/main/java/org/example/WordCRUD.java
@@ -93,7 +93,7 @@ public class WordCRUD implements ICRUD{
         System.out.print("=> 삭제하시겠습니까? (Y/N) : ");
         String ans=kb.nextLine();
         if(ans.equalsIgnoreCase("y")){
-            list.remove(idlist.get(id-1));
+            list.remove((int)idlist.get(id-1));
             System.out.println("단어가 삭제되었습니다.");
         }
         else System.out.println("취소되었습니다.");
